Do not monkey patch scipy functions.
Use newly defined functions instead.
